requirements.txt
requests==2.32.3
clickhouse-connect==0.8.15
python-dotenv==1.0.1
dbt-core==1.8.9
dbt-clickhouse==1.8.8
pytest==8.3.4